MAINT: Moved the tests from `main.rs` to `lib.rs` so that tests could access `log_to_stdout()`
